requirements.txt
livekit>=0.11.0
livekit-agents>=0.8.0
livekit-api>=0.6.0
livekit-plugins-groq>=0.6.0
flask>=2.3.0
groq>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
numpy>=1.24.0
ffmpeg-python>=0.2.0